requirements.txt
dataclasses-json
pip-chill
pygments
pylint
python-dateutil
python-telegram-bot
requests
